requirements.txt
flask==3.0.3
tensorflow==2.13.0
numpy==1.24.3
pillow==10.3.0